Feat: get and return candidate admission status
BREAKING CHANGE

This feature checks for a candidate admission status and returns if admitted or not.
if the candidate is admitted, it returns further, the admitted student activities for the academic session.

It runs asynchronously. When a function that is supposed to take a longer time is invoked,
other non-time taking functions will run without having to waste time.